+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/contentshare/DefaultContentShareController.kt
@@ -5,6 +5,7 @@
 
 package com.amazonaws.services.chime.sdk.meetings.audiovideo.contentshare
 
+import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.LocalVideoConfiguration
 import com.amazonaws.services.chime.sdk.meetings.internal.contentshare.ContentShareVideoClientController
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 
@@ -18,6 +19,12 @@ class DefaultContentShareController(
     override fun startContentShare(source: ContentShareSource) {
         source.videoSource?.let {
             contentShareVideoClientController.startVideoShare(it)
+        }
+    }
+
+    override fun startContentShare(source: ContentShareSource, config: LocalVideoConfiguration) {
+        source.videoSource?.let {
+            contentShareVideoClientController.startVideoShare(it, config)
         }
     }