+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/model/MeetingModel.kt
@@ -53,6 +53,7 @@ class MeetingModel : ViewModel() {
     var isUsingCpuVideoProcessor = false
     var isUsingBackgroundBlur = false
     var isUsingBackgroundReplacement = false
+    var localVideoMaxBitRateKbps = 0
 
     fun updateVideoStatesInCurrentPage() {
         videoStatesInCurrentPage.clear()